Date V1 (unfinished)

+++ ZEngine/include/zengine/Common/Time/Time.hpp
@@ -41,7 +41,7 @@ namespace ze
 		static Time Microseconds(int64_t microseconds) noexcept;
 
 		float asSeconds() const noexcept;
-		float asSecondsInt() const noexcept;
+		int asSecondsInt() const noexcept;
 		int asMilliseconds() const noexcept;
 		int64_t asMicroseconds() const noexcept;
